Source 2 vmap (dmx) format test import

+++ code/platform.h
@@ -6,7 +6,7 @@
 struct ReadFileResult
 {
 	void *contents;
-	u32 size;
+	s64 size;
 };
 
 struct FileInfo
